Ship the vitest base config from core so templates stay standalone
A template directory is the unit that ships: `agent-native create` extracts
it with --strip-components=1, so a config importing the repo-root
`vitest.shared.ts` resolved to nothing in the scaffolded app. For chat and
factory the import sat in `vite.config.ts`, which would have broken dev and
build in every generated app, not just tests.

Move the implementation to `@agent-native/core/vitest-config` — the same
shape templates already use for `@agent-native/core/vite`, and a real
dependency that survives scaffolding. Templates and examples import the
package path; `packages/*` keep the root re-export so they take no
dependency on core.

Verified by extracting templates/tasks the way the CLI does, with nothing
above it: resolves maxWorkers=2. The previous repo-root import fails to load
the config in that same setup.
