+++ .changeset/plenty-jokes-shave.md
@@ -2,9 +2,10 @@
 "@agent-native/scheduling": patch
 "@agent-native/pinpoint": patch
 "@agent-native/dispatch": patch
-"@agent-native/core": patch
+"@agent-native/core": minor
 ---
 
-Cap each vitest suite at a share of the machine through a shared root config so
-concurrent test runs no longer oversubscribe the CPU. Defaults to 25% of cores;
-override with `VITEST_CONCURRENCY`.
+Add `@agent-native/core/vitest-config`, a base vitest config that caps a suite's
+worker pool so concurrent test runs no longer oversubscribe the CPU. Defaults to
+25% of cores; override with `VITEST_CONCURRENCY`. Every template and package
+config merges it in.
