+++ templates/content/vitest.config.ts
@@ -1,9 +1,8 @@
 import path from "node:path";
 
+import baseConfig from "@agent-native/core/vitest-config";
 import react from "@vitejs/plugin-react-swc";
 import { defineConfig, mergeConfig } from "vitest/config";
-
-import baseConfig from "../../vitest.shared";
 
 export default mergeConfig(
   baseConfig,
