+++ templates/mail/vitest.config.ts
@@ -1,8 +1,7 @@
 import path from "node:path";
 
+import baseConfig from "@agent-native/core/vitest-config";
 import { defineConfig, mergeConfig } from "vitest/config";
-
-import baseConfig from "../../vitest.shared";
 
 export default mergeConfig(
   baseConfig,
